feature: add typography resources in the application theme
Use the google provider in order to retrieve the Ubuntu font to use in the application. Add the corresponding dependency in the build.gradle.kts and libs.versions.toml files.
Update all the versions of the libraries to get the latest ones.
Create the dimens.xml file for future dimension references.

Integrates #1