feature: build ui lists for search results, available flights and favorites
Define the layout of the search result suggestions in composable function SearchFlightsScreen, and the layout of the available flights and favorite flights in composables AvailableFlightsScreen and FavoriteFlightsScreen. At the moment, all the composables are collected in file AvailableFlightsScreen.kt, with necessary previews.
Define in dimens.xml and strings.xml the necessary resources to build the ui.

Integrates #3