feature: implement the search bar for searching according to the user input
In the SearchFragment.kt file, define the composable functions for implementing the search bar which displays the results according to the user input.
Rename the AvailableFlightsScreen.kt file to FlightsFragment.kt file, and keep there the composable functions used to build the flights list.
Add a file Utils.kt to collect the composable functions that are used by multiple fragments.
Edit the files dimens.xml and strings.xml according to the needs of the ui implementation.

Integrates #3