feature: put the search bar in the top app bar and separate it from fragments
In build.gradle.kts and libs.versions.toml, add the dependency to import the extended set of icons. Update the Android Gradle Plugin to 8.12.2 version.
Add the MainScreen inside the MainActivity.kt implementation so that the app can be run on an emulator.
In the MainScreen.kt, set the Scaffold composable in order to define the SearchBar, implemented inside the SearchBar.kt file.
Edit the FlightsFragment.kt and SearchFragment.kt accordingly.
Update the strings.xml according to the needs of the ui.

Integrates #3