feature: add view model implementations for MainScreen and fragments
In build.gradle.kts and libs.versions.toml, add the dependency to import the viewmodel-compose dependency with the latest version.
Implement the MainScreenViewModel.kt file to handle the ui state of the main screen, and update the MainScreen.kt and SearchBar.kt composables accordingly, giving functions to the icons and click buttons.
Implement the SearchFragmentViewModel.kt and FlightsFragmentViewModel.kt to handle the state of the SearchFragment.kt and FlightsFragment.kt respectively.

Integrates #3