fix: update the version of the ksp plugin and fix the Room errors
In libs.versions.toml, update the ksp version so that is compatible to the latest kotlin version.
In FavoriteWithAirports.kt, fix the names of the parent columns, matching the database name of the columns.
In FlightSearchDao.kt, add @Transaction to the query involving a relation and the @Delete annotation to the delete function.

Integrates #5