feature: create database and preferences repository and initialize them in application
Implement the two repositories in FlightSearchDatabaseRepository.kt and FlightSearchPreferencesRepository.kt, inserting the functions to get or edit data with Room and preferences data respectively.
In FlightSearchApplication.kt, initialize the two repositories so that can be used as singleton throughout the app.
Import the necessary dependencies and libraries for DataStore preferences in build.gradle.kts and libs.versions.toml files.

Integrates #5