feature: set ui state for SearchFragment
Define the UI state of the SearchFragment.kt by giving the list of airports corresponding to the text search, and changing the list content according to the text provided in the search form.
In FlightSearchApplication.kt, change the definition of the two repository variables using the property delegate lazy.

Integrates #5