feature: set ui state for FlightsFragment
Define the UI state of the FlightsFragment.kt by selecting the list of favorite fragments if the searchText is empty, otherwise return the result of the search.
Update the FlightSearchDao method selectAllAirports() to return all airports when selecting a search result, and add a method to return the list of favorite giving departure and destination. This method is used to check if the flight is favorite or not. Align the methods of the corresponding FlightSearchDatabaseRepository.
Create a new package ui.model which contains the data model of the data for the UI components.

Integrates #5