fix: define the Favorite entity according to the db asset file
Since the db asset file expects two strings, departure_code and destination_code, fix the definition of the Favorite entity so that its model corresponds to the one used in the db asset file.
Update the FlightSearchDao interface and the FlightSearchDatabaseRepository as well, providing a new method to get an Airport based on the iataCode.
As there is no more relationship between the keys of the two entities, delete the FavoriteWithAirports Room relation and delete the relation package.

Integrates #5