feature: adopt a single view model for the entire app
The string to be saved inside the preferences is just the user selection, not the text inputed, as that is only a matter of the UI.
Therefore, delete the view models related to the fragments as they are now useless, and adopt a single view model.
In the view model, define other states for the searchList, the flightsList and also the userSelection, and update the UI Composables accordingly.
Change the name of the preference key to USER_SELECTION, in the FlightSearchPreferencesRepository class.

Integrates #5