feature: handle favorite flights
Generate in the UI an AlertDialog as soon as the user clicks or long-clicks on one of the displayed entries. Add and remove the favorite flights with methods defined in MainScreenViewModel.
In MainScreenViewModel.kt, use the combine function to pilot the flightsList as soon as the userSelection or the favorites list change. Separate the flightsList from the favoritesList and let the UI decide which one to display according to the userSelection.
In Favorite.kt, add a default value to the id property in order to trigger the id auto-generation.
In Utils.kt define the dialog builder and in FlightsFragment.kt define the saveable variables that persist across recompositions.
Use then the state hoisting best practice to define the methods triggered by the alert dialog buttons.

Integrates #5