feature: handle the record audio permission in the app
In order to use the Microphone, allow the app to grant the RECORD_AUDIO permission. Declare the permission in AndroidManifest.xml and import the Accompanist dependency in libs.versions.toml and build.gradle.kts files.
Inside the MainScreen.kt, apply the logic to handle the permission using the rememberPermissionState method for composables. Check the permission only if the user clicks on the Mic icon.
Use the Permissions.kt class to handle a generic type of permission, depending on which that permission must have a rationale or not.
Move SearchBar.kt, SearchFragment.kt, Utils.kt, MainScreen.kt and FlightsFragment.kt inside the screen package and define the viewModel inside the MainActivity.kt file, to enforce the fact that the viewModel has the same lifecycle of the Activity.

Integrates #7