feature: implement the service used when clicking the microphone
Define a service that is going to be used as soon as the user clicks on the Mic icon in the search bar. A service is chosen to handle the recording so that the speech recognizer has its own context; the service is indeed destroyed as soon as the speech result is recorded.
In VoiceRecognitionService.kt, provide the service definition; use VoiceListener.kt as the recognition listener for the speech recognizer defined in the service.
Update the AndroidManifest.xml file by adding the service in the components used by the app.

Integrates #7