feature: display an alert dialog while using the microphone
Implement in Utils.kt an alert dialog that is triggered when the user talks with the speech recognizer activated. Pass the state to the MicUiState.kt and use such state to pilot the UI accordingly.
Define the binding with the VoiceRecognitionService inside the MainScreenViewModel.kt and edit the MicUiState.kt according to the state returned by the speech recognizer.
In MainScreen.kt and MainScreenViewModel.kt, edit the permission handling, making the LaunchedEffect not triggering the Service upon each configuration change, but only when the permission granting is performed for the first time. Delete the Permissions.kt file as it is useless.

Integrates #7